implement support for plugable document save handlers
