+++ backend/de.metas.ui.web.base/src/main/java/de/metas/ui/web/window/descriptor/decorator/IDocumentDecorator.java
@@ -33,6 +33,7 @@ public interface IDocumentDecorator
 	{
 		return ReadOnlyInfo.of(BooleanWithReason.FALSE);
 	}
+
 	@NonNull
 	default BooleanWithReason isDeleteForbidden(@NonNull final TableRecordReference tableRecordReference)
 	{
