+++ backend/de.metas.ui.web.base/src/main/java/de/metas/ui/web/window/model/DocumentsRepository.java
@@ -38,10 +38,14 @@ public interface DocumentsRepository
 
 	OrderedDocumentsList retrieveDocuments(DocumentQuery query, IDocumentChangesCollector changesCollector);
 
-	/** @return document or null */
+	/**
+	 * @return document or null
+	 */
 	Document retrieveDocument(DocumentQuery query, IDocumentChangesCollector changesCollector);
 
-	/** @return document or null */
+	/**
+	 * @return document or null
+	 */
 	default Document retrieveDocumentById(final DocumentEntityDescriptor entityDescriptor, final DocumentId recordId, final IDocumentChangesCollector changesCollector)
 	{
 		return retrieveDocument(DocumentQuery.ofRecordId(entityDescriptor, recordId).setChangesCollector(changesCollector).build(), changesCollector);
@@ -59,17 +63,19 @@ public interface DocumentsRepository
 	 */
 	Document createNewDocument(DocumentEntityDescriptor entityDescriptor, @Nullable final Document parentDocument, final IDocumentChangesCollector changesCollector);
 
-	void refresh(Document document);
+	void refresh(@NonNull Document document);
 
-	SaveResult save(Document document);
+	SaveResult save(@NonNull Document document);
 
-	void delete(Document document);
+	void delete(@NonNull Document document);
 
 	String retrieveVersion(DocumentEntityDescriptor entityDescriptor, int documentIdAsInt);
 
 	int retrieveLastLineNo(DocumentQuery query);
 
-	/** Can be called to verify that this repository belongs with the given {@code entityDescriptor} */
+	/**
+	 * Can be called to verify that this repository belongs with the given {@code entityDescriptor}
+	 */
 	default void assertThisRepository(@NonNull final DocumentEntityDescriptor entityDescriptor)
 	{
 		final DocumentsRepository documentsRepository = entityDescriptor.getDataBinding().getDocumentsRepository();
