+++ backend/de.metas.ui.web.base/src/main/java/de/metas/ui/web/window/model/sql/SqlDocumentsRepository.java
@@ -1,19 +1,13 @@
 package de.metas.ui.web.window.model.sql;
 
 import com.google.common.base.Joiner;
-import com.google.common.collect.ImmutableList;
-import de.metas.cache.model.POCacheSourceModel;
 import de.metas.logging.LogManager;
 import de.metas.ui.web.exceptions.EntityNotFoundException;
 import de.metas.ui.web.session.UserSession;
 import de.metas.ui.web.view.descriptor.SqlAndParams;
-import de.metas.ui.web.window.WindowConstants;
 import de.metas.ui.web.window.controller.DocumentPermissionsHelper;
-import de.metas.ui.web.window.datatypes.DataTypes;
 import de.metas.ui.web.window.datatypes.DocumentId;
-import de.metas.ui.web.window.datatypes.LookupValuesList;
 import de.metas.ui.web.window.descriptor.DocumentEntityDescriptor;
-import de.metas.ui.web.window.descriptor.DocumentFieldDataBindingDescriptor;
 import de.metas.ui.web.window.descriptor.DocumentFieldDescriptor;
 import de.metas.ui.web.window.descriptor.DocumentFieldWidgetType;
 import de.metas.ui.web.window.descriptor.LookupDescriptor;
@@ -27,39 +21,32 @@ import de.metas.ui.web.window.model.DocumentQuery;
 import de.metas.ui.web.window.model.DocumentQueryOrderByList;
 import de.metas.ui.web.window.model.DocumentsRepository;
 import de.metas.ui.web.window.model.IDocumentChangesCollector;
-import de.metas.ui.web.window.model.IDocumentFieldView;
 import de.metas.ui.web.window.model.OrderedDocumentsList;
-import de.metas.ui.web.window.model.lookup.LabelsLookup;
+import de.metas.ui.web.window.model.sql.save.CompositeSaveHandlers;
+import de.metas.ui.web.window.model.sql.save.SaveHandler;
 import de.metas.util.Services;
 import lombok.NonNull;
-import org.adempiere.ad.dao.IQueryBL;
-import org.adempiere.ad.dao.IQueryBuilder;
-import org.adempiere.ad.persistence.TableModelLoader;
+import lombok.RequiredArgsConstructor;
 import org.adempiere.ad.trx.api.ITrx;
 import org.adempiere.ad.trx.api.ITrxManager;
 import org.adempiere.exceptions.AdempiereException;
 import org.adempiere.exceptions.DBException;
 import org.adempiere.exceptions.DBMoreThanOneRecordsFoundException;
-import org.adempiere.model.InterfaceWrapperHelper;
-import org.adempiere.model.PlainContextAware;
 import org.adempiere.service.ClientId;
 import org.adempiere.service.ISysConfigBL;
-import org.adempiere.util.lang.impl.TableRecordReference;
-import org.compiere.model.PO;
-import org.compiere.model.POInfo;
+import org.compiere.model.GridTabVO;
 import org.compiere.util.DB;
 import org.compiere.util.TimeUtil;
+import org.jetbrains.annotations.NotNull;
 import org.slf4j.Logger;
+import org.springframework.stereotype.Component;
 
 import java.sql.PreparedStatement;
 import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.sql.Timestamp;
 import java.time.Instant;
-import java.util.ArrayList;
-import java.util.HashSet;
 import java.util.List;
-import java.util.Set;
 import java.util.concurrent.atomic.AtomicInteger;
 import java.util.function.Function;
 import java.util.stream.Collectors;
@@ -87,15 +74,16 @@ import java.util.stream.Collectors;
  */
 
 /**
- * IMPORTANT: please make sure this is state-less and thread-safe
- *
- * @author metas-dev <dev@metasfresh.com>
+ * @implNote IMPORTANT: please make sure this is thread-safe
  */
+@Component
+@RequiredArgsConstructor
 public final class SqlDocumentsRepository implements DocumentsRepository
 {
-	public static final SqlDocumentsRepository instance = new SqlDocumentsRepository();
-
-	private static final Logger logger = LogManager.getLogger(SqlDocumentsRepository.class);
+	@NonNull private static final Logger logger = LogManager.getLogger(SqlDocumentsRepository.class);
+	@NonNull private final ISysConfigBL sysConfigBL = Services.get(ISysConfigBL.class);
+	@NonNull private final ITrxManager trxManager = Services.get(ITrxManager.class);
+	@NonNull private final CompositeSaveHandlers saveHandlers;
 
 	private static final String VERSION_DEFAULT = "0";
 
@@ -105,19 +93,9 @@ public final class SqlDocumentsRepository implements DocumentsRepository
 	private static final String SYSCONFIG_LoadLimitMax = "webui.documents.LoadLimitMax";
 	private static final int DEFAULT_LoadLimitMax = 300;
 
-	private SqlDocumentsRepository()
-	{
-	}
+	private int getLoadLimitWarn() {return sysConfigBL.getIntValue(SYSCONFIG_LoadLimitWarn, DEFAULT_LoadLimitWarn);}
 
-	private int getLoadLimitWarn()
-	{
-		return Services.get(ISysConfigBL.class).getIntValue(SYSCONFIG_LoadLimitWarn, DEFAULT_LoadLimitWarn);
-	}
-
-	private int getLoadLimitMax()
-	{
-		return Services.get(ISysConfigBL.class).getIntValue(SYSCONFIG_LoadLimitMax, DEFAULT_LoadLimitMax);
-	}
+	private int getLoadLimitMax() {return sysConfigBL.getIntValue(SYSCONFIG_LoadLimitMax, DEFAULT_LoadLimitMax);}
 
 	private static DocumentId retrieveNextDocumentId(final DocumentEntityDescriptor entityDescriptor)
 	{
@@ -397,7 +375,7 @@ public final class SqlDocumentsRepository implements DocumentsRepository
 	}
 
 	@Override
-	public void refresh(final Document document)
+	public void refresh(final @NotNull Document document)
 	{
 		assertThisRepository(document.getEntityDescriptor());
 
@@ -472,75 +450,20 @@ public final class SqlDocumentsRepository implements DocumentsRepository
 	}
 
 	@Override
-	public SaveResult save(final Document document)
+	public SaveResult save(final @NotNull Document document)
 	{
-		Services.get(ITrxManager.class).assertThreadInheritedTrxExists();
+		trxManager.assertThreadInheritedTrxExists();
 		assertThisRepository(document.getEntityDescriptor());
 		DocumentPermissionsHelper.assertCanEdit(document);
 
-		// Runnables to be executed after the PO is saved
-		final List<Runnable> afterSaveRunnables = new ArrayList<>();
-
-		//
-		// Load the PO / Create new PO instance
-		final PO po = retrieveOrCreatePO(document);
-
-		//
-		// Set values to PO
-		final boolean isNew = document.isNew();
-		boolean changes = false;
-		for (final IDocumentFieldView documentField : document.getFieldViews())
-		{
-			if (!isNew && !documentField.hasChangesToSave())
-			{
-				logger.trace("Skip setting PO value because document field has no changes: {}", documentField);
-				continue;
-			}
-
-			if (DocumentFieldWidgetType.Labels == documentField.getWidgetType())
-			{
-				// save labels after PO is saved because we want to make sure it's not new (so we can link to it)
-				afterSaveRunnables.add(() -> saveLabels(document, documentField));
-			}
-
-			if (setPOValue(po, documentField))
-			{
-				changes = true;
-			}
-		}
-
-		//
-		// Save the PO
-		boolean needsRefresh = false;
-		if (changes)
-		{
-			//
-			// Actual save
-			// TODO: advice the PO to not reload after save.
-			InterfaceWrapperHelper.save(po);
-			document.markAsSaved();
-			needsRefresh = true;
-		}
-		else
-		{
-			logger.trace("Skip saving {} because there was no actual change", po);
-		}
-
-		//
-		// Execute after save runnables
-		if (!afterSaveRunnables.isEmpty())
-		{
-			afterSaveRunnables.forEach(Runnable::run);
-			needsRefresh = true;
-		}
+		final SaveHandler.SaveResult handlerResult = saveHandlers.save(document);
 
 		//
 		// Reload the document
 		boolean deleted = false;
-		if (needsRefresh)
+		if (handlerResult.isNeedsRefresh())
 		{
-			final SqlDocumentEntityDataBindingDescriptor dataBinding = document.getEntityDescriptor().getDataBinding(SqlDocumentEntityDataBindingDescriptor.class);
-			final DocumentId idNew = extractDocumentId(po, dataBinding);
+			final DocumentId idNew = handlerResult.getIdNew();
 			final RefreshResult refreshResult = refresh(document, idNew);
 			if (refreshResult == RefreshResult.MISSING)
 			{
@@ -559,282 +482,15 @@ public final class SqlDocumentsRepository implements DocumentsRepository
 		return deleted ? SaveResult.DELETED : SaveResult.SAVED;
 	}
 
-	private DocumentId extractDocumentId(final PO po, SqlDocumentEntityDataBindingDescriptor dataBinding)
-	{
-		if (dataBinding.isSingleKey())
-		{
-			return DocumentId.of(InterfaceWrapperHelper.getId(po));
-		}
-		else
-		{
-			final List<SqlDocumentFieldDataBindingDescriptor> keyFields = dataBinding.getKeyFields();
-			if (keyFields.isEmpty())
-			{
-				throw new AdempiereException("No primary key defined for " + dataBinding.getTableName());
-			}
-
-			final List<Object> keyParts = keyFields.stream()
-					.map(keyField -> po.get_Value(keyField.getColumnName()))
-					.collect(ImmutableList.toImmutableList());
-
-			return DocumentId.ofComposedKeyParts(keyParts);
-		}
-	}
-
-	private PO retrieveOrCreatePO(final Document document)
-	{
-		final SqlDocumentEntityDataBindingDescriptor dataBinding = SqlDocumentEntityDataBindingDescriptor.cast(document.getEntityDescriptor().getDataBinding());
-		final String sqlTableName = dataBinding.getTableName();
-
-		//
-		// Load the PO / Create new PO instance
-		final PO po;
-		if (document.isNew())
-		{
-			po = TableModelLoader.instance.newPO(document.getCtx(), sqlTableName, ITrx.TRXNAME_ThreadInherited);
-		}
-		else if (dataBinding.isSingleKey())
-		{
-			final boolean checkCache = false;
-			po = TableModelLoader.instance.getPO(document.getCtx(), sqlTableName, document.getDocumentIdAsInt(), checkCache, ITrx.TRXNAME_ThreadInherited);
-		}
-		else
-		{
-			po = toQueryBuilder(dataBinding, document.getDocumentId())
-					.create()
-					.firstOnly(PO.class);
-		}
-
-		if (po == null)
-		{
-			throw new DBException("No PO found for " + document);
-		}
-
-		//
-		//
-		po.set_ManualUserAction(document.getWindowNo());
-		InterfaceWrapperHelper.ATTR_ReadOnlyColumnCheckDisabled.setValue(po, true); // allow changing any columns
-
-		//
-		final TableRecordReference rootRecordReference = extractRootRecordReference(document);
-		POCacheSourceModel.setRootRecordReference(po, rootRecordReference);
-
-		return po;
-	}
-
-	private static TableRecordReference extractRootRecordReference(final Document includedDocument)
-	{
-		if (includedDocument.isRootDocument())
-		{
-			return null;
-		}
-
-		final Document rootDocument = includedDocument.getRootDocument();
-		final String rootTableName = rootDocument.getEntityDescriptor().getTableNameOrNull();
-		if (rootTableName == null)
-		{
-			return null;
-		}
-
-		final int rootRecordId = rootDocument.getDocumentId().toIntOr(-1);
-		if (rootRecordId < 0)
-		{
-			return null;
-		}
-
-		return TableRecordReference.of(rootTableName, rootRecordId);
-	}
-
-	private static IQueryBuilder<Object> toQueryBuilder(final SqlDocumentEntityDataBindingDescriptor dataBinding, final DocumentId documentId)
-	{
-		final String tableName = dataBinding.getTableName();
-
-		final List<SqlDocumentFieldDataBindingDescriptor> keyFields = dataBinding.getKeyFields();
-		final int keyFieldsCount = keyFields.size();
-		if (keyFieldsCount == 0)
-		{
-			throw new AdempiereException("No primary key defined for " + tableName);
-		}
-		final List<Object> keyParts = documentId.toComposedKeyParts();
-		if (keyFieldsCount != keyParts.size())
-		{
-			throw new AdempiereException("Invalid documentId '" + documentId + "'. It shall have " + keyFieldsCount + " parts but it has " + keyParts.size());
-		}
-
-		final IQueryBL queryBL = Services.get(IQueryBL.class);
-		final IQueryBuilder<Object> queryBuilder = queryBL.createQueryBuilder(tableName, PlainContextAware.newWithThreadInheritedTrx());
-		for (int i = 0; i < keyFieldsCount; i++)
-		{
-			final SqlDocumentFieldDataBindingDescriptor keyField = keyFields.get(i);
-			final String keyColumnName = keyField.getColumnName();
-			final Object keyValue = convertValueToPO(keyParts.get(i), keyColumnName, keyField.getWidgetType(), keyField.getSqlValueClass());
-			queryBuilder.addEqualsFilter(keyColumnName, keyValue);
-		}
-
-		return queryBuilder;
-	}
-
-	/**
-	 * Sets PO's value from given <code>documentField</code>.
-	 *
-	 * @return true if value was set and really changed
-	 */
-	private static boolean setPOValue(final PO po, final IDocumentFieldView documentField)
-	{
-		final DocumentFieldDataBindingDescriptor dataBinding = documentField.getDescriptor().getDataBinding().orElse(null);
-		if (dataBinding == null)
-		{
-			logger.trace("Skip setting PO's column because it has no databinding: {}", documentField);
-			return false;
-		}
-
-		final POInfo poInfo = po.getPOInfo();
-		final String columnName = dataBinding.getColumnName();
-
-		final int poColumnIndex = poInfo.getColumnIndex(columnName);
-		if (poColumnIndex < 0)
-		{
-			logger.trace("Skip setting PO's column because it's missing: {} -- PO={}", columnName, po);
-			return false;
-		}
-
-		//
-		// Virtual column => skip setting it
-		if (poInfo.isVirtualColumn(poColumnIndex))
-		{
-			logger.trace("Skip setting PO's virtual column: {} -- PO={}", columnName, po);
-			return false; // no change
-		}
-		//
-		// ID
-		else if (poInfo.isKey(poColumnIndex))
-		{
-			final int id = documentField.getValueAsInt(-1);
-			if (id >= 0)
-			{
-				final int idOld = po.get_ValueAsInt(poColumnIndex);
-				if (id == idOld)
-				{
-					logger.trace("Skip setting PO's key column because it's the same as the old value: {} (old={}), PO={}", columnName, idOld, po);
-					return false; // no change
-				}
-
-				final boolean idSet = po.set_ValueNoCheck(columnName, id);
-				if (!idSet)
-				{
-					throw new AdempiereException("Failed setting ID=" + id + " to " + po);
-				}
-
-				logger.trace("Setting PO ID: {}={} -- PO={}", columnName, id, po);
-				return true;
-			}
-			else
-			{
-				logger.trace("Skip setting PO's key column: {} -- PO={}", columnName, po);
-				return false; // no change
-			}
-		}
-		//
-		// Created/Updated columns
-		else if (WindowConstants.FIELDNAMES_CreatedUpdated.contains(columnName))
-		{
-			logger.trace("Skip setting PO's created/updated column: {} -- PO={}", columnName, po);
-			return false; // no change
-		}
-		//
-		// Regular column
-		else
-		{
-			//
-			// Check if value was changed, compared with PO's current value
-			final Object poValue = po.get_Value(poColumnIndex);
-			final Class<?> poValueClass = poInfo.getColumnClass(poColumnIndex);
-			final Object fieldValueConv = convertValueToPO(documentField.getValue(), columnName, documentField.getWidgetType(), poValueClass);
-			if (poFieldValueEqual(fieldValueConv, poValue))
-			{
-				logger.trace("Skip setting PO's column because it was not changed: {}={} (old={}) -- PO={}", columnName, fieldValueConv, poValue, po);
-				return false; // no change
-			}
-
-			//
-			// Check if the field value was changed from when we last queried it
-			if (!po.is_new())
-			{
-				final Object fieldInitialValueConv = convertValueToPO(documentField.getInitialValue(), columnName, documentField.getWidgetType(), poValueClass);
-				if (!poFieldValueEqual(fieldInitialValueConv, poValue))
-				{
-					throw new AdempiereException("Document's field was changed from when we last queried it. Please re-query."
-							+ "\n Document field initial value: " + fieldInitialValueConv
-							+ "\n PO value: " + poValue
-							+ "\n Document field: " + documentField
-							+ "\n PO: " + po);
-				}
-			}
-
-			// TODO: handle not updatable columns... i think we shall set them only if the PO is new
-
-			// NOTE: at this point we shall not do any other validations like "mandatory but null", value min/max range check,
-			// because we shall rely completely on Document level validations and not duplicate the logic here.
-
-			//
-			// Try setting the value
-			final boolean valueSet = po.set_ValueOfColumn(columnName, fieldValueConv);
-			if (!valueSet)
-			{
-				logger.warn("Failed setting PO's column: {}={} (old={}) -- PO={}", columnName, fieldValueConv, poValue, po);
-				return false; // no change
-			}
-
-			logger.trace("Setting PO value: {}={} (old={}) -- PO={}", columnName, fieldValueConv, poValue, po);
-			return true;
-		}
-	}
-
-	/**
-	 * @return true if PO field's values can be considered the same
-	 */
-	private static boolean poFieldValueEqual(final Object value1, final Object value2)
-	{
-		if (value1 == value2)
-		{
-			return true;
-		}
-
-		// If both values are empty we can consider they are equal
-		// (see task https://github.com/metasfresh/metasfresh-webui-api/issues/276)
-		if (isEmptyPOFieldValue(value1) && isEmptyPOFieldValue(value2))
-		{
-			return true;
-		}
-
-		return DataTypes.equals(value1, value2);
-	}
-
-	private static boolean isEmptyPOFieldValue(final Object value)
-	{
-		if (value == null)
-		{
-			return true;
-		}
-		else if (value instanceof String)
-		{
-			return ((String)value).isEmpty();
-		}
-		else
-		{
-			return false;
-		}
-	}
-
 	public static Object convertValueToPO(final Object value, final String columnName, final DocumentFieldWidgetType widgetType, final Class<?> targetClass)
 	{
 		return SqlValueConverters.convertToPOValue(value, columnName, widgetType, targetClass);
 	}
 
 	@Override
-	public void delete(final Document document)
+	public void delete(@NonNull final Document document)
 	{
-		Services.get(ITrxManager.class).assertThreadInheritedTrxExists();
+		trxManager.assertThreadInheritedTrxExists();
 		assertThisRepository(document.getEntityDescriptor());
 		DocumentPermissionsHelper.assertCanEdit(document, UserSession.getCurrentPermissions());
 
@@ -843,9 +499,7 @@ public final class SqlDocumentsRepository implements DocumentsRepository
 			throw new IllegalArgumentException("Cannot delete new document: " + document);
 		}
 
-		final PO po = retrieveOrCreatePO(document);
-
-		InterfaceWrapperHelper.delete(po);
+		saveHandlers.delete(document);
 	}
 
 	@Override
@@ -874,51 +528,8 @@ public final class SqlDocumentsRepository implements DocumentsRepository
 		return DB.getSQLValueEx(ITrx.TRXNAME_ThreadInherited, sql.getSql(), sql.getSqlParams());
 	}
 
-	private static void saveLabels(final Document document, final IDocumentFieldView documentField)
+	public boolean isReadonly(@NonNull final GridTabVO gridTabVO)
 	{
-		final LabelsLookup lookup = LabelsLookup.cast(documentField.getDescriptor().getLookupDescriptor().orElse(null));
-
-		final int linkId = document.getFieldView(lookup.getLinkColumnName()).getValueAsInt(-1);
-		final Set<String> listValuesInDatabase = lookup.retrieveExistingValuesByLinkId(linkId).getKeysAsString();
-
-		final LookupValuesList lookupValuesList = documentField.getValueAs(LookupValuesList.class);
-		final HashSet<String> listValuesToSave = lookupValuesList != null ? new HashSet<>(lookupValuesList.getKeysAsString()) : new HashSet<>();
-
-		//
-		// Delete removed labels
-		{
-			final HashSet<String> listValuesToDelete = new HashSet<>(listValuesInDatabase);
-			listValuesToDelete.removeAll(listValuesToSave);
-			if (!listValuesToDelete.isEmpty())
-			{
-				final int countDeleted = lookup.queryValueRecordsByLinkId(linkId)
-						.addInArrayFilter(lookup.getLabelsValueColumnName(), lookup.normalizeStringIds(listValuesToDelete))
-						.create()
-						.delete();
-				if (countDeleted != listValuesToDelete.size())
-				{
-					logger.warn("Possible issue while deleting labels for linkId={}: listValuesToDelete={}, countDeleted={}", linkId, listValuesToDelete, countDeleted);
-				}
-			}
-		}
-
-		//
-		// Create new labels
-		{
-			final Set<String> listValuesToSaveEffective = new HashSet<>(listValuesToSave);
-			listValuesToSaveEffective.removeAll(listValuesInDatabase);
-			listValuesToSaveEffective.forEach(listValueToSave -> createLabelPORecord(listValueToSave, linkId, lookup));
-		}
-	}
-
-	private static void createLabelPORecord(
-			@NonNull final String listValue,
-			final int linkId,
-			@NonNull final LabelsLookup lookup)
-	{
-		final PO labelPO = TableModelLoader.instance.newPO(lookup.getLabelsTableName());
-		labelPO.set_ValueNoCheck(lookup.getLabelsLinkColumnName(), linkId);
-		labelPO.set_ValueNoCheck(lookup.getLabelsValueColumnName(), listValue);
-		InterfaceWrapperHelper.save(labelPO);
+		return saveHandlers.isReadonly(gridTabVO);
 	}
 }
